error handling changes - moved ErrorCodeType to types package - renamed RestErrorResponse to ErrorResponse - removed status field in ErrorResponse - renamed timestamp field to time in ErrorResponse - removed constructor in ErrorResponse - added builder in ErrorResponse
