+++ src/main/java/com/example/universe/simulator/entityservice/exception/ErrorResponse.java
@@ -1,0 +1,16 @@
+package com.example.universe.simulator.entityservice.exception;
+
+import com.example.universe.simulator.entityservice.types.ErrorCodeType;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.Instant;
+
+@Getter @Setter
+@Builder
+public class ErrorResponse {
+
+    private ErrorCodeType error;
+    private Instant time;
+}
