+++ src/main/java/com/example/universe/simulator/entityservice/validators/MoonDtoValidator.java
@@ -2,7 +2,7 @@ package com.example.universe.simulator.entityservice.validators;
 
 import com.example.universe.simulator.entityservice.dtos.MoonDto;
 import com.example.universe.simulator.entityservice.exception.AppException;
-import com.example.universe.simulator.entityservice.exception.ErrorCodeType;
+import com.example.universe.simulator.entityservice.types.ErrorCodeType;
 import org.springframework.stereotype.Component;
 
 import java.util.Objects;
